Support for 8.5(1) release
