+++ HISTORY.rst
@@ -4,16 +4,23 @@
 History
 =======
 
-1.1.0 (2020-08-21)
+v1.2.0 (2021-2-17)
+---------------------
+
+* Support for 8.5(1) release
+* Some bug fixes and improvements
+
+v1.1.0 (2020-08-21)
 ---------------------
 
 * Support for 8.4(2b) release
 * Many bug fixes and improvements
 
-1.0.1 (2020-05-11)
+v1.0.1 (2020-05-11)
 ---------------------
 
 * Python SDK/API library for Cisco MDS switches
 * PEP8 Compliance
 * Supported modules are switch,devicealias,vsan,zone,zoneset and many more
 * Please check the documentation for more details.
+
