+++ mdssdk/connection_manager/connect_nxapi.py
@@ -57,7 +57,7 @@ class ConnectNxapi(object):
                 payload = dict(
                     jsonrpc=rpc_version,
                     method=method,
-                    params=dict(cmd=command, version=1),
+                    params=dict(cmd=command, version=1.2),
                     id=id_num,
                 )
