+++ mdssdk/portchannel.py
@@ -317,6 +317,8 @@ class PortChannel(Interface):
         )
         out = self.__swobj.show(cmd)
         detailoutput = out["TABLE_port_channel_database"]["ROW_port_channel_database"]
+        if type(detailoutput) is list:
+            detailoutput = detailoutput[0]
         return detailoutput
 
     def __is_pc_present(self):
@@ -345,3 +347,4 @@ class PortChannel(Interface):
         else:
             # There are no PC in the switch
             return False
+
