+++ mdssdk/zone.py
@@ -901,7 +901,13 @@ class Zone(object):
         if out:
             if self.__swobj.is_connection_type_ssh():
                 return out
-            return out["TABLE_zone_status"]["ROW_zone_status"]
+            else:
+                retout = out["TABLE_zone_status"]["ROW_zone_status"]
+                if type(retout) is list:
+                    r = retout[0]
+                else:
+                    r = retout
+            return r
         return out
 
     def _send_zone_show_cmd(self, cmd):
@@ -1016,3 +1022,4 @@ class Zone(object):
                     return
                 else:
                     raise CLIError(cmd, msg)
+
