+++ tests/run_tests.py
@@ -36,19 +36,19 @@ log.info("Starting all tests...")
 
 def get_suite_list(sw):
     suiteList = []
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_device_alias", "test_*.py"))
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_fc", "test_*.py"))
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_port_channel", "test_*.py"))
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_fcns", "test_*.py"))
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_flogi", "test_*.py"))
-    suiteList.append(MyTestLoader(sw).discover("tests.test_switch", "test_*.py"))
-    # suiteList.append(MyTestLoader(sw).discover("tests.test_vsan", "test_*.py"))
-    suiteList.append(MyTestLoader(sw).discover("tests.test_zone", "test_*.py"))
-    suiteList.append(MyTestLoader(sw).discover("tests.test_zoneset", "test_*.py"))
+    #suiteList.append(MyTestLoader(sw).discover("test_device_alias", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_fc", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_port_channel", "test_*.py"))
+    #suiteList.append(MyTestLoader(sw).discover("test_fcns", "test_*.py"))
+    #suiteList.append(MyTestLoader(sw).discover("test_flogi", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_switch", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_vsan", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_zone", "test_*.py"))
+    suiteList.append(MyTestLoader(sw).discover("test_zoneset", "test_*.py"))
     return suiteList
 
 
-for conntype in ["ssh", "https"]:
+for conntype in ["https", "ssh"]:
     with open("switch_details.json", "r+", encoding="utf-8") as f:
         data = json.load(f)
         print(
@@ -77,3 +77,4 @@ hours, rem = divmod(END - START, 3600)
 minutes, seconds = divmod(rem, 60)
 log.info(
     "End of Tests (Took " + ("{:0>1}h:{:0>1}m:{:02.1f}s".format(int(hours), int(minutes), seconds)) + " to complete)")
+
