+++ tests/switch_details.json
@@ -1,5 +1,5 @@
 {
-  "ip_address": "10.126.94.123",
+  "ip_address": "10.126.94.121",
   "username": "admin",
   "password": "nbv!2345",
   "port": 8443,
