+++ tests/test_fc/test_fcattrtransceiver.py
@@ -20,8 +20,9 @@ class TestFcAttrTransceiver(unittest.TestCase):
             k, v = random.choice(list(interfaces.items()))
             if type(v) is Fc:
                 self.fc = v
-                log.debug(k)
-                break
+                if self.fc.transceiver.sfp_present:
+                    log.debug(k)
+                    break
 
     def test_transceiver_read(self):
         self.assertIsNotNone(self.fc.transceiver, "fc.transceiver did not get transceiver objects")
@@ -38,3 +39,4 @@ class TestFcAttrTransceiver(unittest.TestCase):
 
     def tearDown(self) -> None:
         pass
+
