+++ tests/test_fc/vars.py
@@ -30,6 +30,8 @@ status_values = [
     "outofServc",
     "isolated",
     "licenseNotAvail",
+    "init",
+    "offline"
 ]
 
 
@@ -37,3 +39,4 @@ def get_mod_port(fcport):
     mod = fcport.split("/")[0].lstrip("fc")
     port = fcport.split("/")[1]
     return int(mod), int(port)
+
